Create README.md
Project Description: The hiring process is a company's most fundamental and crucial activity. MNCs learn about the major underlying trends in the hiring process here. Trends such as the number of rejections, interviews, job kinds, openings, and so on are critical for a company to consider before recruiting freshers or any other individual. 
As a Data Analyst, our goal is to sift through these trends and extract insights for the hiring department to work with. We work as a lead Data Analyst for a multinational corporation such as Google, and the company has provided us with data records from past hires and has requested us to answer certain questions based on that data. 
We will employ EDA to generate various insights and to respond to the company's questions. Exploratory Data Analysis (EDA) is a method of analysing data that employs visual tools. It is used to detect trends, patterns, or to validate assumptions using statistical summaries and graphical representations.